Rearrange the package contents
