+++ mapping/api.py
@@ -1,3 +1,7 @@
 # flake8: noqa
-
+from .async_loader import AsyncRequest, get_global_async_loader
+from .cacheing_decorators import lfu_cache, lru_cache
+from .http_tile_manager import HTTPTileManager
+from .img_tile_manager import ImageTileManager
+from .mbtile_manager import MBTileManager
 from .utils import get_builtin_mbtiles_path
