+++ mapping/chaco/api.py
@@ -1,3 +1,8 @@
 # flake8: noqa
 from .choropleth_plot import ChoroplethPlot
+try:
+    from .geojson_overlay import GeoJSONOverlay
+except ImportError:
+    pass  # No geojson
 from .map import Map
+from .null_renderer import NullPlotRenderer
