+++ mapping/enable/api.py
@@ -1,12 +1,4 @@
 # flake8: noqa
 from .canvas import MappingCanvas
 from .viewport import MappingViewport
-try:
-    from .geojson_overlay import GeoJSONOverlay
-except ImportError:
-    # No geojson
-    pass
-
-# Tile managers
-from .mbtile_manager import MBTileManager
-from .http_tile_manager import HTTPTileManager
+from .zoom import MappingZoomTool
