+++ mapping/utils.py
@@ -1,11 +1,10 @@
 import os.path as op
 
-import mapping
-
-DATA_DIR = op.join(op.dirname(mapping.__file__), 'data')
+import pkg_resources
 
 
 def get_builtin_mbtiles_path():
     """ Return the location of the included MBTileManager data file.
     """
-    return op.join(DATA_DIR, 'map.mbtiles')
+    return pkg_resources.resource_filename('mapping',
+                                           op.join('data', 'map.mbtiles'))
